requirements.txt
typer==0.21.1
rich==14.3.1
Jinja2==3.1.6
PyYAML==6.0.3
pytest>=8.0.0